+++ lib/frame.dart
@@ -1,5 +1,5 @@
 import 'package:darkstorm_common/nav.dart';
-import 'package:darkstorm_common/top_inherit.dart';
+import 'package:darkstorm_common/top_resources.dart';
 import 'package:flutter/material.dart';
 
 class Frame extends StatefulWidget{