+++ lib/intro.dart
@@ -1,5 +1,5 @@
 import 'package:darkstorm_common/frame_content.dart';
-import 'package:darkstorm_common/top_inherit.dart';
+import 'package:darkstorm_common/top_resources.dart';
 import 'package:flutter/material.dart';
 
 class IntroScreen extends StatefulWidget {
@@ -109,9 +109,8 @@ class IntroPageState extends State<IntroPage> {
   void update() => setState(() {});
 
   @override
-  Widget build(BuildContext context) {
-    print("built");
-    return Column(
+  Widget build(BuildContext context) =>
+    Column(
       mainAxisAlignment: MainAxisAlignment.center,
       crossAxisAlignment: CrossAxisAlignment.center,
       children: [
@@ -131,5 +130,4 @@ class IntroPageState extends State<IntroPage> {
         Container(height: 60)
       ],
     );
-  }
 }